move home page and add translation
