+++ app/components/homepage/Homepage.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { connect } from 'react-redux'
-import { t } from '../utils/translationUtil'
+import { t } from '../../utils/translationUtil'
 import styles from './styles/homepage.css'
 
 const _ = require('lodash')
@@ -10,25 +10,18 @@ const Homepage = ({ addressData }) => {
 
 	return (
 		<div className={styles.main_school_div}>
+			<h4 className={styles.main_header}> {t('home.title')}</h4>
 			<div className={styles.school_left_div}>
 				<div>
-					<p>Your School Address</p>
+					<p>{t('home.addressHeader')}</p>
 					<span>{entry}</span>
-				</div>
-				<div className={styles.page_inner_div}>
-					<p>Your Classrooms</p>
-					<span>Under Development</span>
 				</div>
 			</div>
 
 			<div className={styles.school_right_div}>
 				<div>
-					<p>{t('title')}</p>
-					<span>Under Development</span>
-				</div>
-				<div className={styles.page_inner_div}>
-					<p>Your Subjects</p>
-					<span>Under Development</span>
+					<p>{t('home.resourcesHeader')}</p>
+					<span>Holder</span>
 				</div>
 			</div>
 		</div>
