+++ app/components/homepage/styles/homepage.css
@@ -1,0 +1,21 @@
+.main_header {
+  height: 50px;
+  padding-top: 10px;
+  text-align: left;
+  border-bottom: 1px solid;
+  background: #fdb12e;
+}
+
+.school_left_div {
+  margin: 15px 0 0 30px;
+  float: left;
+}
+
+.school_right_div {
+  margin: 15px 100px 0 0;
+  float: right;
+}
+
+.page_inner_div {
+  margin-top: 20vh;
+}
