+++ app/components/settings/SettingsPage.js
@@ -1,10 +1,15 @@
+/* eslint-disable max-len */
 import React from 'react'
+
+import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
+import { actionCreators } from '../../actions/index'
 import { t } from '../../utils/translationUtil'
 import styles from './styles/settings.css'
 import Address from './Address'
 import GradeFormat from './GradeFormat'
 
-const Settings = () => (
+const Settings = ({ actions }) => (
 	<div className={styles.settings_wrapper}>
 		<h4 className={styles.main_header}>{t('settings.sectionTitle')}</h4>
 		<div className={styles.address_div}>
@@ -13,7 +18,22 @@ const Settings = () => (
 		<div className={styles.gradeFormat_div}>
 			<GradeFormat />
 		</div>
+		<div />
+		<button type="button" className={styles.coffee_div} onClick={actions.buyMeACoffee}>
+			{t('settings.buyMeCoffee')}
+			<img
+				src={`file://${__dirname}/components/settings/helpers/images/BMC.svg`}
+				alt="BMC"
+			/>
+		</button>
 	</div>
 )
 
-export default Settings
+const mapDispatchToProps = dispatch => ({
+	actions: bindActionCreators(actionCreators, dispatch)
+})
+
+export default connect(
+	null,
+	mapDispatchToProps
+)(Settings)
